Create a simpelayout base testing class.
